backend/requirements.txt
fastapi
uvicorn
pymongo
snowflake-connector-python
google-genai
python-dotenv